Fix cyclic reference
